backend/requirements.txt
fastapi
torch
torchvision
uvicorn
Pillow
python-multipart

